Update homework1.py
This program now converts a given number of seconds into hours, minutes and seconds. The main() function requests the user enter the total number of seconds to convert. That number is passed to the three functions which finds the corresponding hours, minutes and remaining seconds and returns them to be printed to screen. The digital clock format isn't perfect and is something I will continue to work on. -Michael Brady